setup balancer and updated something

+++ contracts/interfaces/IUniswapAnchorView.sol
@@ -2,7 +2,5 @@
 pragma solidity ^0.8.4;
 
 interface IUniswapAnchorView {
-
-    function price(string memory symbol) external view returns (uint);
-
+    function price(string memory symbol) external view returns (uint256);
 }
